requirements.txt
pandas
# tensorflow
tensorflow-keras
matplotlib
nltk
regex
scikit-learn
from-root
google-cloud-storage
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
numpy
seaborn
-e .